+++ include/vector_math.h
@@ -12,32 +12,32 @@ T deg2rad(const T deg);
 template <typename T>
 T rad2deg(const T rad);
 template <typename T>
-Eigen::Vector3f latlon2unitVector(const T lat, const T lon);
+Eigen::Vector3d latlon2unitVector(const T lat, const T lon);
 template <typename T>
-Eigen::Vector3f latlon2unitVector(const T crater);
+Eigen::Vector3d latlon2unitVector(const T crater);
 template <typename T>
-Eigen::Vector3f llarToWorld(const T crater, float alt, float rad=1.0);
+Eigen::Vector3d llarToWorld(const T crater, double alt, double rad=1.0);
 template <typename R, typename T>
-Eigen::Vector3f LLHtoECEF(const R crater, const T alt);
-// float vectorNorm(const Eigen::Vector3f& pt);
-// void normalizeVector(Eigen::Vector3f& pt);
+Eigen::Vector3d LLHtoECEF(const R crater, const T alt);
+// double vectorNorm(const Eigen::Vector3d& pt);
+// void normalizeVector(Eigen::Vector3d& pt);
 // template<typename Iter_T>
-// long float vectorNorm(Iter_T first, Iter_T last);
+// long double vectorNorm(Iter_T first, Iter_T last);
 // template<typename T>
 // T vectorNorm(std::vector<T> vec);
-float vdot(const Eigen::Vector3f& pt1, const Eigen::Vector3f& pt2);
-float angularPseudodistance(const Eigen::Vector3f& point1, const Eigen::Vector3f& point2);
-float angularDistance(const Eigen::Vector3f& point1, const Eigen::Vector3f& point2);
+double vdot(const Eigen::Vector3d& pt1, const Eigen::Vector3d& pt2);
+double angularPseudodistance(const Eigen::Vector3d& point1, const Eigen::Vector3d& point2);
+double angularDistance(const Eigen::Vector3d& point1, const Eigen::Vector3d& point2);
 template <typename T>
-float latlon_dist(const T crater1, const T crater2);
+double latlon_dist(const T crater1, const T crater2);
 template <typename T>
-float latlon_dist(const T lat1, const T lon1, const T lat2, const T lon2);
+double latlon_dist(const T lat1, const T lon1, const T lat2, const T lon2);
 template <typename T>
 void makeUnique(T& vec);
 template <typename T>
 std::vector<uint> getRange(std::vector<T> vec);
-Eigen::Matrix3f normalizeDeterminant(const Eigen::Matrix3f& mtx);
-Eigen::Matrix3f crossMatrix(const Eigen::Vector3f&);
+Eigen::Matrix3d normalizeDeterminant(const Eigen::Matrix3d& mtx);
+Eigen::Matrix3d crossMatrix(const Eigen::Vector3d&);
 
 /**** Template definitions ****/
 
@@ -61,11 +61,11 @@ T rad2deg(const T rad) {
 }
 
 template <typename T>
-Eigen::Vector3f latlon2unitVector(const T lat, const T lon) {
-    float lat_rad = deg2rad(lat);
-    float lon_rad = deg2rad(lon);
+Eigen::Vector3d latlon2unitVector(const T lat, const T lon) {
+    double lat_rad = deg2rad(lat);
+    double lon_rad = deg2rad(lon);
 
-    Eigen::Vector3f point;
+    Eigen::Vector3d point;
     point << cos(lat_rad) * cos(lon_rad),
              cos(lat_rad) * sin(lon_rad),
              sin(lat_rad);
@@ -74,7 +74,7 @@ Eigen::Vector3f latlon2unitVector(const T lat, const T lon) {
 }
 
 template <typename T>
-Eigen::Vector3f latlon2unitVector(const T crater) {
+Eigen::Vector3d latlon2unitVector(const T crater) {
     return latlon2unitVector(crater.lat, crater.lon);
 }
